move mandelbulb postprocess to mandelbulb folder
